requirements.txt
fastapi
uvicorn[standard]
azure-ai-formrecognizer
pydantic
python-multipart
python-dotenv